Implemented XML: Added a settings page to allow the user to: - Reload the database, using the .sql script to teardown and rebuild the database - Export the database, exporting the database to be saved as XML. This exports the database schema and the database - Import the database, importing the database restores the current database with an exported XML file. - Close the application NOTE: Do I need to implement editing the XML too, might lose marks for it? Doesn't make much sense to really with my implementation.       Need to think of a way I can incorporate that into the application
Save the PID to a file on execution, and validate that only one instance of the application is running at a time.
Having multiple causes all sorts of bugs so just going to not allow it to happen.
Ignore any XML or database stuff from git to allow a fresh database to be generated when pulled.